or not?

+++ longpoll.php
@@ -5,7 +5,7 @@ $lastId = (int) getLastVote();
  
 $time_wasted = 0;
 $lastIdQuery = ''; 
-$num_rowsbefore= (int) getVotesAfter($timestamp);
+$num_rowsbefore= getVotesAfter($timestamp);
 $num_rows=0;
 if( $num_rows <= 0 ){
    while( $num_rows <= $num_rowsbefore ){
@@ -21,7 +21,7 @@ if( $num_rows <= 0 ){
     
          sleep( 2 );
          $newtimestamp = date('Y-m-d H:i:s');
-         $num_rows= (int) getVotesAfter($newtimestamp);
+         $num_rows= getVotesAfter($newtimestamp);
          $time_wasted += 1;
       }
    }
